requirements.txt
pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0
pillow